37 css arch modes (#38)
* Rename test files to start test_

* Add modes accession and appraisal

* Add tests for accession and appraisal

* Reorganize function calls into new modes

* Add print statement for each mode

The digital archivist has preferred this as a double check that the right mode has been selected when the script's operations are time consuming.

* Update script test for access mode

* Add test data for appraisal mode

* Add script test for appraisal mode

* Add test data for accession mode

* Add script test for accession mode

* Update script test for preservation mode

Not testing what is in appraisal check and appraisal delete logs, which are produced since that function is called for all modes, since they aren't used in this mode.

* Add no file deletion log to access test

* Alphabetize other script outputs

* Explain script modes in top comment

* Explain script modes in README

* Create workflow.md